Project 3 first commit
Added in required functions  plus more gameplay
-Changed navigation to be a linked list system
-Read a list of items and add them to the magic item shop
-Randomize the cost (rounded)
-Give each item a power value based on the cost in order to preserve the
game
-fixed spelling mistakes pointed out
-replaced the showMap() with something much smaller and more dynamic
-Refactored the code to fully work with the linked list and removed all
legacy matrix code
-Added multiple new locales
-Added a jumper bay that the player can enter
-They can use it to go through the gate or explore the planet(to be
added)
-changed gate travel. Require player to find list of gate addresses.
Players than choose what address to dial.
-Therefore direction based travel only works on atlantis, anything
offworld is entirely gate address based
-Lots more to come

+++ src/ListItem.java
@@ -30,10 +30,10 @@ public class ListItem {
         this.next = next;
     }
 
-    public int getCost() {
+    public double getCost() {
         return cost;
     }
-    public void setCost(int cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
 
@@ -61,7 +61,7 @@ public class ListItem {
 
     private String name;
     private String desc;
-    private int cost;
+    private double cost;
     private int power;
     private ListItem next = null;
